requirements.txt
openai==1.70.0
tiktoken==0.9.0
chromadb==1.0.0
langchain==0.3.23
flask
GitPython
python-dotenv  #-->to read credential from .env file
-e .